Add spotify dependencies and auth activity.
Added base spotify auth dependencies and an android library wrapper.
Created an activity to open an authentication flow in a browser window
This is accessed by a single button in the main activity.

Another possible implementation is to not use the web browser, however
as a poc, this is much easier