(Mostly) Fix viewmodel for metrics fragment
Main issue was needing to reference the parent activity, not the fragment itself. The activity is rearranged but the fragment is fully destroyed, therefore so is the viewmodel for it.

The viewmodel appears to be working when rotating the screen but repeated use is bringing up some inconsistencies. Unsure how these occur.

The new vm attributes aren't using mutable data as they don't need to be observed. This would be a nicer way of doing things but I need a button in my app so there's a manual refresh button.