Add generation of recommendation song lists in-app
This is based on a playlist selected in the RecommendationsFragment.
The seed objects are a random selection of 5 of:
 5 of the most frequently
occurring artists in the playlist and 5 randomly selected songs.

This means that there is some randomness in the resulting playlist and
re-running the computation will give a different playlist. Also,
rotating the device will cause a re-run.